Remove explicit dtype conversion in shared expert fusion  (#291)
* add fused_topk

* Remove explict bf16 to fp32 conversion between fused_linear_sigmoid_kernel and _fused_append_shared_experts_gated_kernel


---------

+++ python/sglang/srt/layers/moe/fused_moe_triton/fused_moe_triton_kernels.py
@@ -1186,7 +1186,7 @@ def _fused_append_shared_experts_gated_kernel(
     tl.store(out_ids_ptr + out_ids_row_ptr + offs_k, ids, mask=mask_k)
     tl.store(out_weights_ptr + out_w_row_ptr + offs_k, ws, mask=mask_k)
 
-    shared_w = tl.load(shared_weights_ptr + pid)
+    shared_w = tl.load(shared_weights_ptr + pid).to(ws.dtype)
     offs_s = tl.arange(0, BLOCK_S)
     mask_s = offs_s < S
     shared_ids = tl.cast(N_BASE + offs_s, ids.dtype)
@@ -1218,8 +1218,6 @@ def fused_append_shared_experts_gated(
     out_weights = torch.empty(
         (m, k + s), dtype=topk_weights.dtype, device=topk_weights.device
     )
-
-    shared_weights = shared_weights.to(topk_weights.dtype)
 
     _fused_append_shared_experts_gated_kernel[(m,)](
         topk_ids,
